Kit::Router - Rename Router model && some awesome_print display tweaks
