+++ libraries/kit-router/app/services/kit/router/services/adapters/http/rails/request/import.rb
@@ -13,7 +13,7 @@ module Kit::Router::Services::Adapters::Http::Rails::Request
       _, cookies_ctx = import_rails_cookies(rails_cookies: rails_cookies)
       csrf_token     = rails_controller.session[:_csrf_token] ||= SecureRandom.base64(32)
 
-      router_request = Kit::Router::Models::Request.new({
+      router_request = Kit::Router::Models::RouterRequest.new({
         params: rails_request.params.to_h.symbolize_keys,
         root:   rails_request,
         ip:     rails_request.ip,
